working end to end Kafka -> Flink -> Delta

+++ ch05/applications/flink/flink_iot_stream_writer_delta/src/main/java/io/delta/dldgv2/ch05/utils/DeltaUtils.java
@@ -3,35 +3,16 @@ package io.delta.dldgv2.ch05.utils;
 import org.apache.commons.io.FileUtils;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.util.Optional;
-import java.util.Properties;
 
 public class DeltaUtils {
 
-    public static void prepareDirs(String tablePath) throws IOException {
+    public static void prepareDirs(final String tablePath, boolean deleteIfExists) throws IOException {
         File tableDir = new File(tablePath);
-        if (tableDir.exists()) {
+        if (tableDir.exists() && deleteIfExists) {
             FileUtils.cleanDirectory(tableDir);
         } else {
             tableDir.mkdirs();
         }
-    }
-
-    public static Optional<Properties> loadProperties(final String fileName) throws IOException {
-        FileInputStream fis = null;
-        Properties prop = null;
-        try {
-            fis = new FileInputStream(fileName);
-            prop = new Properties();
-            prop.load(fis);
-        } catch(IOException fnfe) {
-            fnfe.printStackTrace();
-        } finally {
-            if (fis != null) fis.close();
-        }
-        return Optional.of(prop);
     }
 }
